Test: Fix config issue when running specs for json_api_helper from JustMatch
