+++ lib/json_api_helpers/spec/json_api_helpers/serializers/deserializer_spec.rb
@@ -3,6 +3,44 @@
 require 'spec_helper'
 
 RSpec.describe JsonApiHelpers::Serializers::Deserializer do
+  # HACK: Since this spec file can be run from the main JustMatch application
+  #       we need to override the JustMatch test config for these tests in order
+  #       to avoid problems. Both the before and after blocks can be removed if this
+  #       gem is extracted outside of JustMatch
+  before(:each) do
+    @old_deserializer_klass = JsonApiHelpers.config.deserializer_klass
+    @old_params_klass = JsonApiHelpers.config.params_klass
+
+    JsonApiHelpers.configure do |config|
+      config.deserializer_klass = Class.new do
+        def self.jsonapi_parse(json_api_hash)
+          json_api_hash.dig('data', 'attributes')
+        end
+      end
+
+      config.params_klass = Class.new do
+        def initialize(hash)
+          @hash = hash
+        end
+
+        def [](name)
+          @hash[name.to_s]
+        end
+
+        def to_h
+          @hash
+        end
+      end
+    end
+  end
+
+  after(:each) do
+    JsonApiHelpers.configure do |config|
+      config.deserializer_klass = @old_deserializer_klass
+      config.params_klass = @old_params_klass
+    end
+  end
+
   let(:params) do
     {
       'data' => {
